Set the lot owner in the HouseData packet
The second dword in the HRMT chunk in the HouseData packet is the
Lot Owner field. Setting this field places a special Lot Owner icon on
the lower right of the Sim's icon in the UI at the bottom of the screen,
and also displays "Owner - " in the tooltip when you hover over the
Sim's icon.

HRMT chunk contents (updated as of 2018-02-24):
* Unknown - 4 bytes; should be 1
* Property Leader / Lot Owner - 4 bytes; in the NoNetwork lot, this
  defaults to kNewAvatarID (0xffff1111).